Saki's new gun effects, various new sounds and some QoL changes. Take #2!
Saki's projectiles as well as the behavior of some neutral b, up throw and final smash aspects updated, new sounds added. On top of that, fixing an issue with his pummel not damaging and fixing dsmash's animation appearance as a bonus.
Side credit to Sato for helping me with the coding aspects.

Co-Authored-By: Sato <62856711+Sato1999@users.noreply.github.com>